Add flag to hoist instructions
This fixes issue #6654

Only hoist instructions that are optimized by prepareFuncForForwardDiff.
Do this by adding a flag tryToHoist to IRSimplificationOptions and set
this to true only if called from prepareFuncForForwardDiff, then only
hoist if the flag is set. Additionally, do not hoist loops if they only
have a single trivial iteration.